refactor: directory pages renamed screens
