+++ nym-vpn-app/src/screens/Error.tsx
