+++ nym-vpn-app/src/screens/home/ConnectionBadge.tsx
