+++ nym-vpn-app/src/screens/home/ModeDetailsDialog.tsx
