+++ nym-vpn-app/src/screens/settings/InfoData.tsx
