+++ nym-vpn-app/src/screens/settings/appearance/display/Display.tsx
