+++ nym-vpn-app/src/screens/settings/appearance/display/UiScaler.tsx
