+++ nym-vpn-app/src/screens/settings/legal/Legal.tsx
