+++ nym-vpn-app/src/screens/settings/legal/LegalRouteIndex.tsx
