+++ nym-vpn-app/src/screens/settings/legal/index.ts
