+++ nym-vpn-app/src/screens/settings/legal/licenses/LicenseDetails.tsx
