+++ nym-vpn-app/src/screens/settings/legal/licenses/index.ts
